Refactor, using JSON snapshot for Expore tab previews
